+++ tests/scenario/test_traffic_sign_interpreter.py
@@ -10,10 +10,10 @@ from commonroad.scenario.traffic_sign import (
     TrafficSignElement,
     TrafficSignIDZamunda,
 )
-from commonroad.scenario.traffic_sign_interpreter import TrafficSigInterpreter
+from commonroad.scenario.traffic_sign_interpreter import TrafficSignInterpreter
 
 
-class TestTrafficSigInterpreter(unittest.TestCase):
+class TestTrafficSignInterpreter(unittest.TestCase):
     def setUp(self):
         lanelet_one = Lanelet(
             left_vertices=np.array([[0.0, 0.0], [1.0, 0.0], [2, 0]]),
@@ -76,7 +76,7 @@ class TestTrafficSigInterpreter(unittest.TestCase):
         lanelet_network.add_traffic_sign(traffic_sign_three, {102})
         lanelet_network.add_traffic_sign(traffic_sign_four, {103})
 
-        self.interpreter = TrafficSigInterpreter(SupportedTrafficSignCountry.ZAMUNDA, lanelet_network)
+        self.interpreter = TrafficSignInterpreter(SupportedTrafficSignCountry.ZAMUNDA, lanelet_network)
 
     def test_speed_limit(self):
         self.assertEqual(20, self.interpreter.speed_limit(frozenset({100, 101, 102})))